pyproject.toml
[project]
name = "todo-list-api-auth-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "mysqlclient>=2.2.8",
]
